requirements.txt
beautifulsoup4==4.14.2
gspread==6.2.0
python-dotenv==1.2.1
requests==2.32.5
